Pull back the be folder from the original HaroldXL repo

+++ be/src/main/java/com/smokingcessation/repository/PlanTasksFreeRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface PlanTasksFreeRepository extends JpaRepository<PlanTasksFree, Long> {
+public interface PlanTasksFreeRepository extends JpaRepository<PlanTasksFree, Integer> {
     List<PlanTasksFree> findByUser_UserId(Integer userId);
     boolean existsByUserAndTaskDay(User user, LocalDate taskDay);
     List<PlanTasksFree> findByUserAndTaskDay(User user, LocalDate taskDay);
